Update steps / screenshot info
The screenshot, where it was displayed, is incorrect.  I had a customer contact me today regarding this specifically because his output doesn't show that way.  I worked it in my test lab where I had not yet prepared for PTA, and indeed the Domains [] value was blank (in addition, the command you're instructed to run is Get-AzureADSSOStatus | ConvertFrom-Json, but the output shows the result of the command without being piped through ConvertFrom-Json).
I updated with just the normal text values of a 'before' installation, and moved the screenshot to the end where it should be displayed.